Changed how countdown works
If the event hasn't started yet, then the header says "starts in" and it counts down to the start of the event.
If the event is ongoing, the header says "ends in" and it counts down to the end of the event
If the event is over, the header says "is over" and the timer reads all 0's.
Also changed the color of the timer to yellow for readability.